// frontend/src/api/models/event-type-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Call Booking Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DurationMinutes } from './duration-minutes';

export interface EventTypeCreateRequest {
    'title': string;
    'description': string;
    'durationMinutes': DurationMinutes;
    'availableFrom': string;
    'availableTo': string;
    'color': string;
}